# DocSwagger.yaml
swagger: '2.0'
info:
  description: >-
    This page is the documentation of QrCode's API.
  version: 1.0.0
  title: Documentation API QrCode
host: 'localhost:8080'
basePath: /
schemes:
  - http
paths:
  /qrcodes:
    get:
      tags:
      - "qrcodes"
      summary: Get all QrCodes of database
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
          
    post:
      tags:
      - "qrcodes"
      summary: "Add new QrCode"
      description: ""
      operationId: "addQrCode"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: ""
      responses:
        405:
          description: "Invalid input"
      security:
      - api_key: []
  
  "/qrcode/{uid}":
    get:
      tags:
        - qrcode
      summary: Get one QrCode by UID
      description: Retourne une seule equipe
      operationId: getQrCodeById
      produces:
        - application/json
      parameters:
        - name: uid
          in: path
          required: true
          type: string
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid UID supplied
        '404':
          description: QrCode not found
      security:
        - api_key: []
    delete:
      tags:
      - "qrcode"
      summary: "Deletes a QrCode"
      description: ""
      operationId: "deleteQrCode"
      produces:
      - "application/json"
      parameters:
      - name: "uid"
        in: "path"
        description: "QrCode uid to delete"
        required: true
        type: "string"
        format: "string"
      responses:
        400:
          description: "Invalid UID supplied"
        404:
          description: "QrCode not found"
      security:
      - api_key: []
  "/qrcode":   
    put:
      tags:
      - "qrcode"
      summary: "Update an existing QrCode"
      description: ""
      operationId: "updatePet"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: ""
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "QrCode not found"
        405:
          description: "Validation exception"
      security:
      - api_key: []
    
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header